src/main.cc: Improve the seed of srand
A bug that generates the same "random" name of the
device each time on first boot has been detected.
It has been observed on Raspberry Pi 3 and other
devices. The following fix improves the seed of
srand to avoid this bug. For details have a look at:

PRO-3542 aktualizr autoprovisioning always selects the same "random" name

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>