Initialize PRNG from /dev/random, fix unaligned read
Embedded systems often have a lack of randomness, especially early in the boot
sequence. To reduce the risk that repeated reads from /dev/urandom give the
same result, using /dev/random instead.  This will block until there is enough
randomness available.

Also fix an issue where a char* pointer is converted to a int* pointer.  While
this is safe on Intel architectures, it is not necessarily OK on ARM and MIPS,
where an unaligned read will fault.  Instead this allocates storage for an
unsigned int, and casts the pointer to a char*, which is guaranteed to work.