Tidy up Base64 handling in utils
sotarviclient.cc had a duplicate of code in Utils::fromBase64, so use
Utils::fromBase64 directly in sotarivclient.cc.

The public interface of utils.h was also pulling in various headers that are
only needed for the implementation--include these in the utils.cc
implementation instead.

Finally, add some unit tests of the Base64 functionality.