main.cc: Handle missing configuration file
Gracefully terminate the application if the
configuration file is missing or if there is another
runtime unhandled exception.

Although the Aktualizr recipe in Yocto/OE layer
meta-updater deploys sota.toml as a read-only file,
it will be missing if the user forgets to specify
variable SOTA_PACKED_CREDENTIALS. This fix ensures
that in this case Aktualizr will not crash.

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>