Restore -Wall -Wextra -Werror, etc.
Some careful manipulation was necessary to make jsoncpp and rvi_lib
still work. Furthermore, a few exceptions were necessary for gtest.

This naturally required several code changes. Most are insignificant and
the rest should be unnoticeable. The only possible exception is that I
got rid of a crypto test case that is no longer necessary with compiling
with these flags.