Catch problems identified with -Wpedantic.
New option PEDANTIC_WARNINGS will enable -Wpedantic as well as the
warnings that are caused by gtest. This is useful for identifying issues
locally without requiring everyone to suffer from these flags. Combining
this with WARNING_AS_ERROR will probably cause the build fail unless
everything is just right.

For the record, gcc 6.0.0 defaults to std=gnu++14. Before that, it was
gnu++98. Currently, if not building with BUILD_GENIVI, we don't specify
a language standard, so the default is used.