Remove 'Pinkel' from the wordlist
Pinkel is a type of smoked sausage from northern Germany, however it is similar
to the verb 'pinkeln', which is slang for 'to use the gentleman's lavatory'.
At the risk of offending our friends in Lower Saxony, we should remove this
from the list in order to save the blushes of those who are not acquainted with
this dish.