Refactor TLS config into a class.
The goal is to allow the cert/ca/pkey files to be specified as an
absolute path or a relative path (with certificates_directory as the
root as has been hitherto assumed). This is necessary for installing a
root CA correctly with implicit_writer. It also cleans up a lot of code
that conconates paths by now doing it in one place.