Move uptane serial tests to their own file.
Unfortunately this has a rather minimal effect on total run time. My
goal here was to reduce the total time running tests and make
uptane_test a little shorter. The latter goal is easily obtained, but
the former is difficult. I'm not really seeing the improvement I was
hoping for.

Also remove a couple PROJECT_WORKING_DIRECTORY statements that weren't
necessary.