Switch to C++11 by default
Now is the time to move Aktualizr to C++11.  At the moment the core of the product compiles against C++98, although we require C++11 for some of the support utilities.

C++11 came in gcc around version 4.7, and Debian oldstable is currently at gcc 4.9, so I hope that we should be able to assume C++11 now.