Fix test concurrency bug (and minor OSTree logic bug).
Main problem was that it was wrong to have multiple tests checking for
the presence of the sysroot and trying to construct it if not found.
Solved by doing it once as a prerequisite for running tests.

Also found an obscure bug in our OSTree logic, but it only is revealed
by doing several things wrong in sequence.