Provide safer temporary files and directories
As mentioned in PRO-4721 and PRO-4722, TemporaryFile and
TemporaryDirectory were somewhat unsafe.

To solve this, we create a global temporary directory root for the
lifetime of the application, with restrictive permissions. The directory
is created in the OS' canonical temporary directory so that aktualizr
can eventually run under a lesser privileged user.