Improve resilience of example_flasher logic.
Previously it assumed if /var/sota existed it was good to use. Now it
actually checks for read-/writability before moving forward. This is
obviously still liable to fail if the filesystem changes before we
actually use it, but this is just an example that is also used for
testing.