Integerate get_garage_sign into CMake. Untar tuf-cli package in the script.
Define GARAGE_SIGN_VERSION to control the version to download and
package, e.g. cmake ../ -DGARAGE_SIGN_VERSION=cli-0.3.0-5-g5908997.tgz

The default is the version whose name sorts highest, which should be the
greatest version number. If the layout or content of the tuf-cli
releases page changes substantially, this may break.