Fix some library dependency caching problems.
Previously, if you had built with certain features enabled and then
turned them off, CMake could complain about not finding the libraries
anymore. The fix was to wipe the build directory. Now it should just
work. Thanks to Laurent for helping track this down.

I also updated the list of dependencies in the README.