ostree: Make compatible with 2016.15
The CI container uses Xenial and the 2016.15 version of ostree. This
version does not include the ostree_async_progress_get function. It
also doesn't include the "caught-error" and "scanning" items. However,
its easy enough to get the "scanning" status in a way that fails
gracefully with the old api and works with the new one.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>