uptane: Change logging for unknown keys
After doing a key rotation described here:

 https://docs.atsgarage.com/prod/rotating-signing-keys.html

We started seeing the LOG_INFO message. It seemed a bit concerning
to people until we dug into it more. I think this message is probably
more useful as a LOG_DEBUG. Additionally, if we are going to warn
about a key I think we ought to show it.

Signed-off-by: Andy Doan <andy@opensourcefoundries.com>