Use asn1c generated serializers
When we originally decided to use our own serialization code, we hoped this
would mean that we could reuse our existing domain objects, but in practice we
haven't been able to do that, and are going bytes -> message -> domain object
anyway in many cases.  If we are going to have an intermediate 'message'
object, then it would seem a good idea to use an auto-generated one.

We currently have code that uses tag values and dynamic casts to deal with
discriminated unions of message types, and this would be replaced with the
code generated by asn1c, which is a enum + struct--so no worse, really.

Debugging will be made easier, because we can dump a human and machine readable
representation of the message using XER.

Finally, since BER messages are self-delimiting, and asn_decode() supports
reading partial messages from a stream, we can switch to using a single TCP
connection for the entire protocol, rather start a TCP connection per message,
meaning that the secondary won't have to open a connection to the primary any
more.

This commit only moves secondary discovery over to the the system. Once this is
in we can move the rest of the protocol over.