Only write the important parts of the config with implicit_writer.
We no longer need to dump the entire thing, since we can use the
systemd-style config construction now. This simplifies and solves
several things.

We now use config dumping for printing the final configuration to the
log, though, so I changed the tests to use that logic to make sure the
output looks reasonable.

I also did a bit of cleanup of aktualizr_secondary_config to get closer
to the main aktualizr config style.