Fix enum default logic for storage and package manager types.
Previously, because a default was provided as a string, if two or more
configs were provided, and the last one did not specify the enum type,
it would fall back to the string default (instead of the previously set
value). This has been fixed by further extending the template magic.

Also fix some typos.