Move config dir checking to a common function.
Also use it for the Config(vector<path>) constructor, which
cert_provider uses. Previously, a nonexistent config would be silently
ignored. Now it is as least not silent.

The idea is that if config files/dirs specified directly on the
commandline do not exist, you should tell the user, because that
probably matters.