Fix metadata migration bugs
This was introducted in #e458e89011, and causes migration from v5 to v6 to fail
in all cases. There are a couple of fixes, firstly the repo_types table doesn't
match the C++ enum values, and secondly the INSERT INTO meta_migrate lines were
not consistent with the table definitions (although the contents of meta_types
and the C++ do match).

I considered adding a new migration to v7 to fix this, but this won't help the
cases where people have already migrated from v5 to v6. Any devices in that
situation are now bricked, so there is no way to save them.

There is a now a test case in sqlstorage_test.cc to check that any future
migrations keep a valid root.json.

Any devices created with the old v6 should continue to work, except they will
have funny values in the repo_types table. Fortunately we don't use this at the
moment, and since we never 'released' this version officially hopefully any
devices with the old repo_types table will get reset before we depend on it.