Fix enum default logic for CryptoSource and KeyType.
Previously, because a default was provided as a string, if two or more
configs were provided, and the last one did not specify the enum type,
it would fall back to the string default (instead of the previously set
value). This has been fixed by further extending the template magic.

This was fixed for storage and package manager types in
1ac08de47c6d09d0d407f0d1945df1ef2b6764ae, but unfortunately, this was
not caught at that time.