SendDeviceData also for --running-mode=check.
This will send a lot of redundant data every time aktualizr runs with
that flag, but that might be reasonable anyway if the data changed since
the last run.

This probably isn't a perfect solution, but since there is no specific
requirement at present, I'd rather keep it simple while sending the
messages that were previously being skipped in manual runs.