Use new config for FS to SQL migration
Our current migration procedure doesn't solve the issue of paths to tls
creds changed from the default... Since we instantiate `FSStorage` from
a blank config, it will only find them in that case.

Modify the procedure so that it's possible to change the file names in
the new (SQL) config. Not perfect but makes the migration possible!