Clean the docker mess a bit
Do not use `.nop11`, `noostree` docker images anymore, let's settle with
per-distro dockerfiles and trust our CMake for not linking with extra
libs. We are reasonably confident that `cmake -DBUILD_OSTREE=off` does
not try to link with Ostree and keeping all these files up to date has
been costly compared to the benefits.

Also, move everything to a dedicated directory to clean up the root of
the project.