Refactor to create new aktualizr_test.
Roughly based around the idea of eventsinterpreter_test, using some of
the existing uptane_tests to cover similar ground. Basically, we need
tests to cover the expected event flow of the various running modes,
which is similar to what the old eventsinterpreter_test was doing. There
are parts still missing and a lot of cleanup to do.