Unit tests for OSTreeObject class.
Really only covers the constructor, MakeTestRequest, and Upload. These
are the components that are easiest to verify and have the most external
interaction. Testing CurlDone or the various OSTree parsing bits might
be interesting, but it is quite difficult to decouple from other parts
of the code, and the whole of it is already covered by higher-level
tests. I'll take suggestsions for improvements, but I think this is a
good start.