Investigate TODOs in external actions section.
Leave comments in the tests that correspond to the items in the actions
list to make it easier to connect them in the future (and easier to
understand the goals of some of the tests). I didn't do this for
aktualizr_test, as it is fairly high-level and covers a bunch of things,
although it's kind of the only place to do some things.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>