Only sleep as long as it takes for the server to be ready.
Lots of tests sleep for 3 or 4 seconds in the expectation that the test
server will be ready in that time. However, our CI regularly fails when
that assumption is not true. So instead of increasing the time, just
poll the server for readiness and sleep only as long as necessary.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>