Tests: rework fetcher test
Fetcher test currently uses `large_interrupted` file for pause/resume.
With `large_interrupted` server aborts after transferring half of a file
(1000 bytes). This should lead to the `The target's calculated hash did
not match the hash in the metadata` error and after that, the upper layer
may choose to retry.
However, currently, the test succeeds, because while committing the first
half to the db, fetcher gets `pause` event, which makes it retry
silently after `resume` is called.
This behavior itself is not a problem, the problem is that it is also
correct for the test to fail, if `pause` comes after the first chunk
is committed.
To solve this:
* Use `large_file`
* Launch download in a separate thread, so we can timeout,
  currently after 20 sec.
* Pause the fetcher after a certain percent is downloaded,
  rather than after 200 ms.

Signed-off-by: Eugene Smirnov <evgenii.smirnov@here.com>