Fetcher: use condition variable for pausing
Using a mutex to pause the fetcher is not only non-idiomatic, but it is also
an undefined behavior, as the thread which called `pause` may terminate
and/or resume might get called from another thread.
See [https://en.cppreference.com/w/cpp/thread/mutex] and
[https://en.cppreference.com/w/cpp/thread/mutex/unlock].
It seems to work now, but it can easily break anytime.

Also there was a race condition between `fetchVerifyTarget()`
or `DownloadHandler()` and `setPause()`, which should be solved
with this commit:
   http->download()
                                               setPause(true)
   DownloadHandler() {
   ...
   return written_size + 1
                                               setPause(false)
   http->download() == CURLE_WRITE_ERROR &&
   pause_ == false
   throw OversizedTarget

Signed-off-by: Eugene Smirnov <evgenii.smirnov@here.com>