Get rid of running modes.
aktualizr still has an input parameter (--run-mode) that does the same
thing, but this is no longer wired directly into libaktualizr; it exists
purely in the wrapper outside it. This should be much cleaner and
simpler to use.

Note that once upon a time I advocated for all aktualizr commandline
options to be also represented in the config. I no longer believe this
to be a good idea for things that only matter outside of libaktualizr.
An internal "run mode" for libaktualizr does not make sense. It should
be driven by the code outside of the API.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>