Tests: fetcher resume after death
A test scenario for the use-case when aktualizr process gets killed while being
in a `pause` state. When the fetcher is called again, it's expected to start
the download from the point it was paused on.
Currently, a delay of 10 seconds between `pause` call and killing the process
is needed, as we have no reliable way to know when all background operations
are completed, so we can resume later.

Signed-off-by: Eugene Smirnov <evgenii.smirnov@here.com>