Remove kRejectAll test, because it is useless.
It was rendered useless at the latest in
b763ae6b7dc4caa4ebbd8e7deaeb0079e51f1a74, which is the commit that
stopped using the specially-crafted test database.

Metadata is handled significantly differently now, and this test was not
doing anything else interesting to warrant keeping it around.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>