Fix issue with never-ending install retries
As backend sends empty director targets in some unexpected situations,
we've implemented a client-side mitigation in
02deeb5cc45e2a1fac4f6e3eb515a396cd14c1c3.

One unwanted side-effect is that after a failed installation, we'll keep
taking into account the targets containing the new version, hence
retrying installations in a loop until the backend asks to install
something new.

Second fix is to drop director targets after installation end, until
we've come to an agreement on a proper solution on both sides.

Signed-off-by: Laurent Bonnans <laurent.bonnans@here.com>