OTA-2606: Renaming Provisioning Terminology
This is part of the wider task of implementing renaming.
I replace topics that I consider to be duplicates of what
is in the docs repo. I indicated this in a comment at the
top of the file. The replacements contain the new
terminology. I also updated the terminology in the
"non-duplicate" files.

Signed-off-by: Merlin Carter <merlin.carter@here.com>
Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>