Replace cert auth in garage tools with mutual TLS support.
The old method required seperate client cert, key, and root CA in the
credentials archive. It has been unused as far as I can tell for at
least a year, if not two or more. The one test for it has been broken
for a long time.

That method has been replaced with mutual TLS support, which is what the
backend is transitioning to. OAuth2 support will eventually get dropped
on the server, but in the meantime we still need to support it for some
transitional period.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>