aktualizr-lite: Use current target for status
Now that we are tracking the installed target properly, we can use
getCurrent to make this easier and more informative.

This also uses a reverse iterator for the OstreeManager:getCurrent
method. In the event you have two targets with the same hash, you
should probably match on the newest one rather than the oldest one.

Signed-off-by: Andy Doan <andy@foundries.io>