Improve Target metadata matching logic.
Added a check for HW IDs. The Director sends a map of serials to HWIDs,
but the Image repo sends us a vector of HWIDs. If the map and vector
match, the comparison succeeds. Otherwise, we expect that every HWID in
the map can be found in the vector. Added tests for all of these cases.

Also added explanations for the rest of the fields we use for why we
don't check them.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>