Recheck the Uptane metadata before downloading or installing a target.
uptaneIteration() is now exclusively used during the check for updates,
and it actually fetches the necessary metadata.

uptaneOfflineIteration() is now exclusively used during downloading and
installing to verify the integrity of the stored metadata. If a user
then attempts to install a target that doesn't match the stored
metadata, it will be rejected.

That stipulation required a lot of reworking of the Uptane tests,
because several of them installed fake packages as a shortcut. That no
longer works, so I had to expand several of the tests quite a bit. As a
result, several of these tests are now rather redundant. We should
re-evaluate the test cases in uptane_test and aktualizr_test and
determine which are still meaningful and unique.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>