Actually check downloaded hash during verification.
checkTargetFile() just checks that the stored metadata is consistent. We
need to go further and actually recheck the hash of the downloaded data.

Tested at a low level and high level for the fake package manager. Only
tested at a high level for OSTree, and because of the limitations of the
mocked sysroot, it's not as thorough as it could be.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>