Revert source code to 2019.5 state
The 2019.5-docs branch was an attempt to clean up the branch so that
it only had the docs-relevant commits in it. This involved a pretty
massive cherry-pick, and many of the committed commits included
changes to both code and docs. This last commit just takes out a
shotgun and reverts everything except for the docs/ directory to the
way it was when 2019.5 was released.

Signed-off-by: Jon Oster <jon.oster@here.com>