Clean up and link in meta-updater testing and advanced usage pages.
These two pages are still good and contain unique information, so I
simply cleaned them up and added them to the navigation. I reframed the
usage page as "advanced usage".

I removed the OSTree section from the usage page, as it had been
untouched for years and was mostly bogus by this point. We have better
ways of demonstrating it at this point anyway.

Co-Authored-By: Jon Oster <tkfu@users.noreply.github.com>
Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>