Use latest endpoint for /system_info
rvi core is deprecated and not in use any more, so going forward all
references to it should be removed. device-registry handles this api
now, and the last versions of core forward the data to device-registry.

Device gw should proxy /core/system_info to
api/v1/devices/:uuid/system_info, which is handled by device registry.

Signed-off-by: Simão Mata <simao.mata@here.com>