OstreeManager: refactor hash check to prevent bogus warning.
By calling getCurrent() in finalizeInstall(), an error was printed that
the stored and deployed hashes did not match. That is expected in that
case. Besides, there was no reason to check the stored hash at that time
at all, so I've refactored the code to only get the hash at that time
and thus only check what we really care about.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>