Fix logging issue from bug #1477
This is subtle so I'm keeping what's really two changes as one commit:

The issue is that DockerAppManger's constructor needs to create a
"fake fetcher" that's not really used but required for constructing
member variables and they don't have access to the global Config. Only
the PackageManagerConfig. Calling the `Config` default constructor
results in a call to `postUpdateValues` that then calls
`logger_set_threshold` with the default logging level, INFO.

This commit updates the Fetch constructor to only use the values that
it needs from the globabl config: repo-server and director-server. The
DockerAppMgr constructor then doesn't to create a fake global config.

Signed-off-by: Andy Doan <andy@foundries.io>