Reject all targets if one doesn't match.
Previously, checkUpdates() returned any targets that were successfully
matched before one that did not. That was inconsistent, it conflicted
with the Uptane standard, and it caused an unnecessary download call in
UptaneCycle(), which correctly rejected all targets.

Also inform the backend of the installation failure in this case,
although the installation wasn't even attempted.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>