aktualizr-lite: Include primary ECU for target download
With the introduction of commit
 fead283be7813688f9f69037d2c879a9857157c5

SotaUptaneClient::downloadImage now requires the Target to have
EcuSerial information that would normally come from Uptane. When running
aktualizr-lite, the Target is taken directly from the TUF targets and
does not include this information.

This change allows aktualizr-lite to:

1) Set primary ECU information in the SotaUptaneClient
This change isn't too bad (splitting logic into smaller pieces).
However, it would make more sense if this was a private method.

2) Set Ecu information on the Uptane::Target object.

Signed-off-by: Andy Doan <andy@foundries.io>