dockerappmanager: Improve "fetch" logic
The current fetch logic grabs the docker app file, but it doesn't pull
down the container(s) for the app until it runs "docker-compose start".
This isn't optimal or what people are expecting the "fetch" step to do.
This change pulls down the containers during the fetch phase. It has
a side-effect in that the "verify" step is now basically redundant since
the docker app file must be rendered so that the fetch can be performed.

Signed-off-by: Andy Doan <andy@foundries.io>