Rename "Images" to "Image repo" and standardize capitalization.
The Uptane standard refers to an "Image repository" but never an "images
repository". I've caught most trivial references and corrected them.
I've also searched for other common examples of deviation from the
naming norms of Uptane, mostly in terms of capitalization. This might
seem trivial but should improve clarity when trying to read our code
and follow the standard. For guidance on this subject, see:
https://github.com/uptane/uptane-standard#style-guide

Note that it is particularly confusing to talk about Targets (roles and
metadata) vs. targets (updates for ECUs). Capitalization helps
distinguish.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>