Improve logging of binary downloads.
Previously, if a binary download was initiated, you'd see a slightly
confusing and daunting message like "No file 'xyz' with
matched hash in the database". Now that only appears if the file was
expected to be present and was not. Also added some similar messages
(mostly at debug level) for other failure and edge cases.

Suggested-by: Xavier Gully <xavier.gully@here.com>
Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>