Add SecondaryProvider class.
This allows a much cleaner, simpler, and more easily extensible
interface for providing the Secondary implementations on the Primary
with the data they require.

Metadata is now fetched through this instead of using an inflexible
struct. This only changes the interface but nothing substantial about
the actual implementations. Work still has to be done to actually
support delegations, but this already simplifies things for partial
verification.

Reading the target image and Treehub credentials through this allows us
to get rid of ImageReaderProvider and the aktualizr API call for the
credentials.

The one questionable aspect is adding the init() call to the
SecondaryInterface. It should be harmless, and may even be useful in the
future, but it could be done differently.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>