Match new updates with the installed versions with MatchTarget.
This is the same function used for matching Targets from the Director
and Image repo. Notably, it also matches hashes and length.

Functionally, this shouldn't change much in practice, but if a Target is
received in the Director Targets metadata that matches the filename of
the currently installed version but *not* the length and/or hashes, we
throw an error, since that is unexpected. This is covered by a test.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>