Create new type MetaBundle to replace RawMetaPack.
RawMetaPack had a fixed size, which makes it impractical for using with
delegations. This way we can use whatever metadata we have, and whenever
we get around to supporting delegations for Secondaries or
Timestamp/Snapshot from the Director, it won't be a big deal to add them
in as needed.

Greater flexibility comes with the cost of verbosity, but I don't think
it's too bad.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>