Improve error reporting in IP Secondary ASN.1 protocol.
Now putMetadata, sendFirmware, and install all can return a return code
and a description. This should get forwarded all the way back to the
backend now.

Note that some of other requests (like getInfor and getManifest) do not
yet use anything like that. I didn't see a need, but I'm open to
discussion.

Protocol negotiation is now only done during putMetadata, so that's the
only thing that gets retried if a Secondary is still using the old v1
protocol messages. This can still be improved a bit.

I also explicitly labeled some of the messages that are now deprecated
as part of the first version of the protocol.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>