Only send device data if it has changed.
We now stored hashes of the hardware/system information, list of
installed packages, network information, and configuration in the
database. The hardware and network data were previously stored locally
in SotaUptaneClient, but this persists over restarts, so it saves even
more bandwidth.

Note that at least on my laptop, one field in the output of lshw changes
every time I call it, so the hardware info gets sent every time.
Thankfully, I could not reproduce that problem on qemu.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>