Fix flaky test by waiting for Secondary to start up.
The Primary is now more efficient and I guess the Secondary needs some
extra time to catch up now. I wish there was a better solution here.

I fixed a typo and tried to make the expiry tests a bit more consistent
with each other since that was originally what I thought might be a
cause of the failure.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>