Test IP Secondary RPC failure processing.
I refactored the RPC tests substantially but functionally only added
another test that verifies that errors reported by the Secondary are
correctly interpreted on the Primary (both result code and description).

We already cover that these errors are reported from the Secondary
implementations on the Primary back to the server via the installation
report thanks for the VirtualSecondary failure injection tests.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>