Test adding, removing, and replacing IP Secondaries.
The logic for matching addresses to ECUs is not trivial, so it's worth
being extra careful. Also test replacing an ECU that reuses the same
address and port and that we correctly track an ECU that changes its
address or port.

Also note some the timeouts (and some ordering issues) that were
necessary to get some of these tests to pass. These are more reasons why
a proper connection manager would be nice to have; things like this
really shouldn't be necessary.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>