Always read the device ID from the cert if available.
For device credential provisioning, it is still a fatal error if the
cert is unavailable. For shared cred prov, we revert to the old behavior
of generating a random pet name if the cert is unavailable.

This is generally only observable if you attempt to replace a Primary
ECU, which requires reusing the same device certificate.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>