Remove shared cred p12 from the archive after provisioning.
It should never be needed again and presents a potential security risk
if left available for an attacker who gains access to the device.

Also remove treehub.json. It shouldn't be there in the first place, but
in the distant past it was often accidentally left there.

Improved some RAII with archive structs, too.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>