ProvisionMode is now actually configurable.
The default is still to decide based on whether provision_path is empty,
but you can also make your preferences explicit. There is also a new
option: kSharedCredReuse, which will skip the removal of the shared
credential p12 from the archive. This is intended to be used only for
testing.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>