Refactor Virtual Secondary config writing for tests.
Previously we were writing a separate config file for each Secondary,
despite that the Primary can only use one at a time. Now we create only
one and append to it if it already exists. We still store the
Secondary-specific files in their own subdirectories, but the exact
layout is a bit different. Several tests needed adjustment to these
changes, but functionally, nothing changed.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>