Mark Boost as a 'System' library, to avoid linting it.
If Aktualizr is being built against a copy of Boost that is installed from the
package manager in /usr, then gcc already knows that it is a system library,
and doesn't trigger any warnings for those headers. However if cmake finds a
copy of Boost somewhere else, then this logic doesn't happen and the build
fails because Boost isn't warning-clean with the agressive set of flags in
Aktualizr.

Marking it as 'SYSTEM' tells cmake to tell gcc it is a system library, and the
warnings are suppressed as normal.

https://cmake.org/cmake/help/latest/command/include_directories.html
Signed-off-by: Phil Wise <phil@phil-wise.com>